talks/2011: copy in the lexical scanning talk from its old location
It resides in my code.google.com personal repository, but that
will go away and anyway it belongs here, a place that did not
exist when the talk was given.

The only edits are formatting for the present format, an added
line about Go 1 allowing goroutines in init, and a file rename
to avoid a git gofmt precommit.

Change-Id: I4c2f2c2d44fef9cb574ef4f7223d9cb0c7b15060
Reviewed-on: https://go-review.googlesource.com/12839
Reviewed-by: Andrew Gerrand <adg@golang.org>